requirements.txt
allure-pytest==2.13.5
pytest==8.2.0
selenium==4.20.0